@@ -1,13 +1,12 @@
 plugins {
     id 'java'
     id 'maven-publish'
-//    id 'com.github.johnrengelman.shadow' version '7.1.2'
     id 'java-library'
-    id "io.ebean" version "13.14.0" apply(false)
+    id "io.ebean" version "13.17.2" apply(false)
 }
 group 'org.start2do'
 
-version '2.0.5-SNAPSHOT'
+version '3.0.0-SNAPSHOT'
 allprojects {
     apply {
         plugin 'java'
@@ -29,19 +28,18 @@ allprojects {
         lombokVersion = '1.18.22'
         jasperreportsVersion = '6.19.1'
         mapstructVersion = "1.4.2.Final"
-        ebeanVersion = "13.14.0"
+        ebeanVersion = "13.17.2-jakarta"
         vertxVersion = '4.3.4'
+        springBootVersion = '3.0.6'
+        springVersion = '6.0.3'
+        springCloudVersion = '2022.0.0'
+        springCloudStarterVersion = '3.1.7'
+        springAliCloudVersion = '2022.0.0.0-RC2'
         curVersion = project.version == 'unspecified' ? this.rootProject.version : project.version
-        jdkVersion = ["ebean-common"        : JavaVersion.VERSION_11,
-                      "spring-common"       : JavaVersion.VERSION_11,
-                      "spring-security-api" : JavaVersion.VERSION_11,
-                      "spring-security-core": JavaVersion.VERSION_11,
-                      "http-log-sdk": JavaVersion.VERSION_11,
-                      "spring-business"     : JavaVersion.VERSION_11,
-        ]
+        jdkVersion = []
     }
     compileJava {
-        def curJdkVersion = JavaVersion.VERSION_1_8
+        def curJdkVersion = JavaVersion.VERSION_17
         for (final def j in jdkVersion) {
             if (j.key == project.name) {
                 curJdkVersion = j.value
@@ -55,10 +53,12 @@ allprojects {
         compileOnly("org.projectlombok:lombok:$lombokVersion")
         annotationProcessor("org.projectlombok:lombok:$lombokVersion")
         compileOnly("org.mapstruct:mapstruct:$mapstructVersion")
+        compileOnly("org.slf4j:slf4j-api:1.7.36")
         annotationProcessor("org.mapstruct:mapstruct-processor:$mapstructVersion")
         annotationProcessor("io.ebean:querybean-generator:$ebeanVersion")
-        annotationProcessor group: 'org.springframework.boot', name: 'spring-boot-configuration-processor', version: '2.6.7'
-        compileOnly("org.springframework.boot:spring-boot-autoconfigure:2.6.7")
+        annotationProcessor group: 'org.springframework.boot', name: 'spring-boot-configuration-processor', version: springBootVersion
+        compileOnly group: 'org.springframework.boot', name: 'spring-boot-autoconfigure', version: springBootVersion
+
     }
 
     test {
