@@ -3,9 +3,9 @@ package org.start2do.entity;
 import io.ebean.Model;
 import io.ebean.annotation.DbName;
 import java.time.LocalDateTime;
-import javax.persistence.Column;
-import javax.persistence.EmbeddedId;
-import javax.persistence.Entity;
+import jakarta.persistence.Column;
+import jakarta.persistence.EmbeddedId;
+import jakarta.persistence.Entity;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
