@@ -2,7 +2,7 @@ package org.start2do.dto;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
 import java.util.UUID;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
