@@ -1,7 +1,8 @@
 package org.start2do;
 
-import javax.annotation.PostConstruct;
+import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -10,7 +11,9 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.start2do.config.KaptchaConfig;
 import org.start2do.util.JwtTokenUtil;
+import org.start2do.util.StringUtils;
 
+@Slf4j
 @Configuration(proxyBeanMethods = false)
 @ComponentScan("org.start2do")
 @Import({Start2doSecurityConfig.class, KaptchaConfig.class})
@@ -26,6 +29,12 @@ public class AutoScanConfig {
 
     @PostConstruct
     public void init() {
-        JwtTokenUtil.SECRET = start2doSecurityConfig.getSecret();
+        String secret = start2doSecurityConfig.getSecret();
+        if (StringUtils.isEmpty(secret)) {
+            secret = org.start2do.util.JwtTokenUtil.genKey();
+            start2doSecurityConfig.setSecret(secret);
+            log.info("生成密钥:{}", secret);
+        }
+        JwtTokenUtil.SECRET = secret;
     }
 }
