@@ -1,7 +1,7 @@
 package org.start2do.dto.req.login;
 
 import java.io.Serializable;
-import javax.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotEmpty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
