@@ -1,11 +1,11 @@
 package org.start2do.util;
 
+import jakarta.validation.Validation;
+import jakarta.validation.Validator;
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.ValidatorFactory;
 import java.util.Set;
 import java.util.StringJoiner;
-import javax.validation.ConstraintViolation;
-import javax.validation.Validation;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
 
 public final class BeanValidatorUtil {
 
