@@ -4,23 +4,35 @@ import io.ebean.Database;
 import io.ebean.DatabaseFactory;
 import io.ebean.config.CurrentUserProvider;
 import io.ebean.config.DatabaseConfig;
+import io.ebean.migration.MigrationConfig;
+import io.ebean.migration.MigrationRunner;
 import io.ebean.spring.txn.SpringJdbcTransactionManager;
 import javax.sql.DataSource;
-import org.springframework.beans.factory.annotation.Value;
+import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.start2do.ebean.id_generators.UUIDStrIdGenerator;
 
+@Import(EbeanConfig.class)
 @Configuration
-@ConditionalOnProperty(name = "spring.datasource.username")
+@ConditionalOnProperty(name = "spring.datasource.url")
+@RequiredArgsConstructor
 public class EbeanBeanAutoConfiguration {
 
-    @Value("${migration:false}")
-    private static boolean migration = false;
+    private final EbeanConfig ebeanConfig;
 
+    private void migration(DataSource dataSource) {
+        if (!ebeanConfig.isMigration()) {
+            return;
+        }
+        MigrationConfig config = new MigrationConfig();
+        MigrationRunner runner = new MigrationRunner(config);
+        runner.run(dataSource);
+    }
 
     @Bean
     @ConditionalOnMissingBean(DatabaseConfig.class)
@@ -29,11 +41,14 @@ public class EbeanBeanAutoConfiguration {
         config.loadFromProperties();
         config.add(new UUIDStrIdGenerator());
         config.setCurrentUserProvider(currentUserProvider);
-        config.setRunMigration(migration);
+        config.setRunMigration(ebeanConfig.isMigration());
         config.setDataSource(dataSource);
         config.setDdlRun(false);
         config.setExternalTransactionManager(new SpringJdbcTransactionManager());
         config.setDdlCreateOnly(false);
+        if (ebeanConfig.isMigration()) {
+            migration(dataSource);
+        }
         return config;
     }
 
@@ -43,7 +58,7 @@ public class EbeanBeanAutoConfiguration {
         return () -> "not set";
     }
 
-    @Bean(name = "Database")
+    @Bean
     @ConditionalOnMissingBean(Database.class)
     @ConditionalOnBean(value = {DatabaseConfig.class})
     public io.ebean.Database database(DatabaseConfig config) {
