@@ -1,26 +1,28 @@
 package org.start2do.util.spring;
 
+import jakarta.annotation.PostConstruct;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import javax.annotation.PostConstruct;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.Accessors;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
 
 
 @Setter
 @Getter
 @Accessors(chain = true)
 @NoArgsConstructor
+@Configuration
 @ConfigurationProperties(prefix = "start2do.log")
 public class LogAopConfig {
 
@@ -40,5 +42,7 @@ public class LogAopConfig {
             ByteArrayOutputStream.class
         ));
     }
+
+
 
 }
